pyproject.toml
[tool.poetry]
name = "test-poetry"
version = "0.1.0"
description = ""
authors = ["hustchihieu143 <chihieusky@gmail.com>"]
readme = "README.md"
packages = [{include = "test_poetry"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.1"
config = "^0.5.1"
dotenv = "^0.0.5"
fastapi-utils = "^0.2.1"
fastapi-mail = "^1.1.5"
Flask-Mail = "^0.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
